refactor(environment): promote environment summary surfaces
